refactor!: moving the JsonCompiler from Ark to CLI include directory
